Added most of the slash commands

+++ Random/repeat.py
@@ -1,3 +1,7 @@
-async def repeat(ctx, message):
-    await ctx.message.delete()
-    await ctx.channel.send(message)
+async def repeat(ctx, message, type):
+    if type == 'normal':
+        await ctx.message.delete()
+        await ctx.channel.send(message)
+    elif type == 'slash':
+        await ctx.respond('The message has been repeated.',ephemeral=True)
+        await ctx.channel.send(message)
